python-agent/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
jsonschema==4.25.1
PyJWT==2.10.1
cryptography==45.0.6
openai==1.102.0
